Move the DefaultCache implementation to Impl
Move the DefaultCache implementation to DefaultCacheImpl class

Resolves: OLPEDGE-1587
